requirements.txt
Pillow
tk
streamlit